don't copy inexisting file

+++ tests/scripts/fusionforge-build-and-test-rpm.sh
@@ -71,7 +71,7 @@ then
 	scp -p /tmp/timedhosts.txt root@$HOST:/var/cache/yum/timedhosts.txt
 fi
 ssh root@$HOST "yum install -y fusionforge fusionforge-plugin-scmsvn fusionforge-plugin-online_help fusionforge-plugin-extratabs fusionforge-plugin-ldapextauth fusionforge-plugin-scmgit fusionforge-plugin-blocks"
-scp -p root@$HOST:/var/cache/yum/timedhosts.txt /tmp/timedhosts.txt
+[ ! -e /tmp/timedhosts.txt ] || scp -p root@$HOST:/var/cache/yum/timedhosts.txt /tmp/timedhosts.txt
 ssh root@$HOST "cd /usr/share/tests/func; CONFIGURED=true CONFIG_PHP=config.php.buildbot DB_NAME=$DB_NAME php db_reload.php"
 ssh root@$HOST "su - postgres -c \"pg_dump -Fc $DB_NAME\" > /root/dump"
 # Install a fake sendmail to catch all outgoing emails.
